sebelum revisi flowmap

+++ resources/views/engineer/lemburpengajuan.blade.php
@@ -43,7 +43,7 @@
                                     Lokasi
                                 </th>
                                 <th scope="col" class="px-2 py-3 ">
-                                    Detail Pengajuan
+                                    Lama Kegiatan
                                 </th>
                                 <th scope="col" class="px-2 py-3 ">
                                     Engineer Status
@@ -82,7 +82,7 @@
                                     {{ $keg->kegiatan }}
                                 </td>
                                 <td class="px-2 py-2 border-r">
-                                    {{ $keg->lokasi }}
+                                    {{ $keg->lama_kegiatan }}
                                 </td>
                                 <td class="px-2 py-2 border-r">
                                     <div class="flex justify-center m-5">
@@ -193,7 +193,7 @@
                                     Kegiatan
                                 </th>
                                 <th scope="col" class="px-2 py-3 ">
-                                    Lokasi
+                                    Lama Kegiatan
                                 </th>
                                 <th scope="col" class="px-2 py-3 ">
                                     Detail Pengajuan
@@ -233,7 +233,7 @@
                                     {{ $keku->kegiatan }}
                                 </td>
                                 <td class="px-2 py-2 border-r">
-                                    {{ $keku->lokasi }}
+                                    {{ $keku->lama_kegiatan }}
                                 </td>
                                 <td class="px-2 py-2 border-r">
                                     <div class="flex justify-center m-2">
@@ -298,11 +298,17 @@
                                 </td>
                                 <td class="px-2 py-2">
                                     <div class="flex justify-center">
-                                        <a href="/engineer/datapengajuan/delete/{{ $keku->id_kegiatan }}">
-                                            <svg class="w-6 h-6 text-red-400 hover:text-red-500" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" width="24" height="24" fill="currentColor" viewBox="0 0 24 24">
-                                                <path fill-rule="evenodd" d="M8.586 2.586A2 2 0 0 1 10 2h4a2 2 0 0 1 2 2v2h3a1 1 0 1 1 0 2v12a2 2 0 0 1-2 2H7a2 2 0 0 1-2-2V8a1 1 0 0 1 0-2h3V4a2 2 0 0 1 .586-1.414ZM10 6h4V4h-4v2Zm1 4a1 1 0 1 0-2 0v8a1 1 0 1 0 2 0v-8Zm4 0a1 1 0 1 0-2 0v8a1 1 0 1 0 2 0v-8Z" clip-rule="evenodd"/>
-                                            </svg>                               
-                                        </a>
+                                        @if($keku->statacc_engineer === 'ditolak' || $keku->statacc_manager === 'ditolak')
+                                            <a href="/engineer/datapengajuan/delete/{{ $keku->id_kegiatan }}">
+                                                <svg class="w-6 h-6 text-red-400 hover:text-red-500" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" width="24" height="24" fill="currentColor" viewBox="0 0 24 24">
+                                                    <path fill-rule="evenodd" d="M8.586 2.586A2 2 0 0 1 10 2h4a2 2 0 0 1 2 2v2h3a1 1 0 1 1 0 2v12a2 2 0 0 1-2 2H7a2 2 0 0 1-2-2V8a1 1 0 0 1 0-2h3V4a2 2 0 0 1 .586-1.414ZM10 6h4V4h-4v2Zm1 4a1 1 0 1 0-2 0v8a1 1 0 1 0 2 0v-8Zm4 0a1 1 0 1 0-2 0v8a1 1 0 1 0 2 0v-8Z" clip-rule="evenodd"/>
+                                                </svg>                               
+                                            </a>
+                                        @else
+                                                <svg class="w-6 h-6 text-gray-500" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" width="24" height="24" fill="currentColor" viewBox="0 0 24 24">
+                                                    <path fill-rule="evenodd" d="M8.586 2.586A2 2 0 0 1 10 2h4a2 2 0 0 1 2 2v2h3a1 1 0 1 1 0 2v12a2 2 0 0 1-2 2H7a2 2 0 0 1-2-2V8a1 1 0 0 1 0-2h3V4a2 2 0 0 1 .586-1.414ZM10 6h4V4h-4v2Zm1 4a1 1 0 1 0-2 0v8a1 1 0 1 0 2 0v-8Zm4 0a1 1 0 1 0-2 0v8a1 1 0 1 0 2 0v-8Z" clip-rule="evenodd"/>
+                                                </svg>                               
+                                        @endif
                                     </div>
                                 </td>
                             </tr>
